+++ ghost/core/core/server/services/mentions/MentionCreatedEvent.js
