+++ ghost/core/test/unit/server/services/mentions/MentionDiscoveryService.test.js
@@ -1,10 +1,12 @@
-const MentionDiscoveryService = require('../lib/MentionDiscoveryService');
 const sinon = require('sinon');
-// non-standard to use externalRequest here, but this is required for the overrides in the library, which we want to test for security reasons in combination with the package
-const externalRequest = require('../../core/core/server/lib/request-external.js');
 const dnsPromises = require('dns').promises;
 const assert = require('assert/strict');
 const nock = require('nock');
+
+// non-standard to use externalRequest here, but this is required for the overrides in the library, which we want to test for security reasons in combination with the package
+const externalRequest = require('../../../../../core/server/lib/request-external.js');
+
+const MentionDiscoveryService = require('../../../../../core/server/services/mentions/MentionDiscoveryService');
 
 describe('MentionDiscoveryService', function () {
     const service = new MentionDiscoveryService({externalRequest});
