--- ghost/webmentions/lib/webmentions.js
@@ -1,6 +1,0 @@
-module.exports.InMemoryMentionRepository = require('./InMemoryMentionRepository');
-module.exports.MentionsAPI = require('./MentionsAPI');
-module.exports.MentionDiscoveryService = require('./MentionDiscoveryService');
-module.exports.Mention = require('./Mention');
-module.exports.MentionSendingService = require('./MentionSendingService');
-module.exports.MentionCreatedEvent = require('./MentionCreatedEvent');
