--- ghost/webmentions/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@tryghost/webmentions",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/webmentions",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --check-coverage --100  --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "bson-objectid": "2.0.4",
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "nock": "13.5.6",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/errors": "1.3.7",
-    "@tryghost/logging": "2.4.21",
-    "cheerio": "0.22.0"
-  }
-}
